requirements.txt
goprocam==4.2.0
Pillow==9.4.0
opencv-python==4.8.1.78
numpy==1.24.2
pygame==2.1.2



